Fix issue #30
