+++ herringbone/__init__.py
@@ -1,5 +1,5 @@
 from .env_core import (
     Piece, Board, MDP, TransitionMatrix, Episode, Action, Policy, Algorithm, 
-    PolicyIteration, ValueIteration, load_map, Trajectory, MonteCarloController, MonteCarloPredictor, EpsilonGreedyPolicy
+    PolicyIteration, ValueIteration, load_map, Color, Trajectory, MonteCarloController, MonteCarloPredictor, EpsilonGreedyPolicy
 )
 print("imported herringbone without any errors :)")
