+++ herringbone/env_core/utils/__init__.py
@@ -1,1 +1,2 @@
 from .map_loader import load_map
+from .color import Color
